@@ -9,79 +9,76 @@ import (
 	"os"
 	"os/signal"
 
-	"github.com/bufbuild/protovalidate-go"
+	"connectrpc.com/connect"
+	"connectrpc.com/validate"
 	"github.com/caarlos0/env/v11"
-	"github.com/pcriv/mancala/internal/mancala"
-	"github.com/pcriv/mancala/proto/protoconnect"
 	"github.com/redis/go-redis/v9"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
 	"golang.org/x/sync/errgroup"
 
+	"github.com/pcriv/mancala/internal/mancala"
 	redisstore "github.com/pcriv/mancala/internal/store/redis"
+	"github.com/pcriv/mancala/proto/protoconnect"
 )
 
 type envConfig struct {
-	Env      string `env:"ENV" envDefault:"local"`
-	LogLevel string `env:"LOG_LEVEL" envDefault:"debug"`
+	Env      string `env:"ENV"                envDefault:"local"`
+	LogLevel string `env:"LOG_LEVEL"          envDefault:"debug"`
 	RedisURL string `env:"REDIS_URL,required"`
-	Address  string `env:"ADDRESS" envDefault:":50051"`
+	Address  string `env:"ADDRESS"            envDefault:":50051"`
 }
 
 func main() {
-	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, os.Kill)
-	defer cancel()
-
-	err := run(ctx)
+	logger, err := run()
 	if err != nil && !errors.Is(err, context.Canceled) {
-		slog.Error("error running application", slog.String("error", err.Error()))
+		logger.Error("error running application", slog.String("error", err.Error()))
 		os.Exit(1)
 	}
 
-	slog.Info("closing server gracefully")
+	logger.Info("closing server gracefully")
 }
 
-func run(ctx context.Context) error {
+func run() (*slog.Logger, error) {
+	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
+	ctx, cancel := signal.NotifyContext(context.Background(), os.Interrupt, os.Kill)
+	defer cancel()
+
 	cfg := envConfig{}
 	if err := env.Parse(&cfg); err != nil {
-		return fmt.Errorf("unable to parse config: %w", err)
+		return logger, fmt.Errorf("unable to parse config: %w", err)
 	}
-
-	slog.SetDefault(slog.New(slog.NewJSONHandler(os.Stdout, nil)))
 
 	redisClient, err := newRedisClient(ctx, cfg.RedisURL)
 	if err != nil {
-		return err
+		return logger, err
 	}
 	gameStore := redisstore.NewGameStore(redisClient)
 
-	validator, err := protovalidate.New()
-	if err != nil {
-		return fmt.Errorf("failed to create validator: %w", err)
-	}
-
 	path, handler := protoconnect.NewServiceHandler(
 		handler{
-			service:   mancala.NewService(gameStore),
-			validator: validator,
+			service: mancala.NewService(gameStore),
 		},
+		connect.WithInterceptors(
+			validate.NewInterceptor(),
+		),
 	)
 
 	mux := http.NewServeMux()
 	mux.Handle(path, handler)
 
 	// Use h2c so we can serve HTTP/2 without TLS.
-	srv := http.Server{
-		Addr:    cfg.Address,
-		Handler: h2c.NewHandler(mux, &http2.Server{}),
+	svr := http.Server{
+		Addr:              cfg.Address,
+		Handler:           h2c.NewHandler(mux, &http2.Server{}),
+		ReadHeaderTimeout: 0,
 	}
 
 	g, ctx := errgroup.WithContext(ctx)
-
 	g.Go(func() error {
-		slog.Info("starting connect server on address", slog.String("address", cfg.Address))
+		logger.Info("starting connect server on address", slog.String("address", cfg.Address))
 
-		if err := srv.ListenAndServe(); err != nil {
+		if svrErr := svr.ListenAndServe(); svrErr != nil {
 			return fmt.Errorf("failed to listen and serve connect service: %w", err)
 		}
 
@@ -91,14 +88,14 @@ func run(ctx context.Context) error {
 	g.Go(func() error {
 		<-ctx.Done()
 
-		if err := srv.Close(); err != nil {
+		if svrErr := svr.Close(); svrErr != nil {
 			return fmt.Errorf("failed to close server: %w", err)
 		}
 
 		return nil
 	})
 
-	return g.Wait()
+	return logger, g.Wait()
 }
 
 func newRedisClient(ctx context.Context, url string) (*redis.Client, error) {
