@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.0
+// 	protoc-gen-go v1.36.11
 // 	protoc        (unknown)
 // source: mancala.proto
 
@@ -15,6 +15,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -603,103 +604,64 @@ func (x *Player) GetScore() int64 {
 
 var File_mancala_proto protoreflect.FileDescriptor
 
-var file_mancala_proto_rawDesc = []byte{
-	0x0a, 0x0d, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
-	0x07, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x1a, 0x1b, 0x62, 0x75, 0x66, 0x2f, 0x76, 0x61,
-	0x6c, 0x69, 0x64, 0x61, 0x74, 0x65, 0x2f, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x61, 0x74, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x59, 0x0a, 0x11, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x47,
-	0x61, 0x6d, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x21, 0x0a, 0x07, 0x70, 0x6c,
-	0x61, 0x79, 0x65, 0x72, 0x31, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x42, 0x07, 0xba, 0x48, 0x04,
-	0x72, 0x02, 0x10, 0x01, 0x52, 0x07, 0x70, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x31, 0x12, 0x21, 0x0a,
-	0x07, 0x70, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x32, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x42, 0x07,
-	0xba, 0x48, 0x04, 0x72, 0x02, 0x10, 0x01, 0x52, 0x07, 0x70, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x32,
-	0x22, 0x46, 0x0a, 0x12, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x47, 0x61, 0x6d, 0x65, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x30, 0x0a, 0x0c, 0x63, 0x72, 0x65, 0x61, 0x74, 0x65,
-	0x64, 0x5f, 0x67, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0d, 0x2e, 0x6d,
-	0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x47, 0x61, 0x6d, 0x65, 0x52, 0x0b, 0x63, 0x72, 0x65,
-	0x61, 0x74, 0x65, 0x64, 0x47, 0x61, 0x6d, 0x65, 0x22, 0x21, 0x0a, 0x0f, 0x46, 0x69, 0x6e, 0x64,
-	0x47, 0x61, 0x6d, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x0e, 0x0a, 0x02, 0x69,
-	0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x22, 0x35, 0x0a, 0x10, 0x46,
-	0x69, 0x6e, 0x64, 0x47, 0x61, 0x6d, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12,
-	0x21, 0x0a, 0x04, 0x67, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0d, 0x2e,
-	0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x47, 0x61, 0x6d, 0x65, 0x52, 0x04, 0x67, 0x61,
-	0x6d, 0x65, 0x22, 0x5e, 0x0a, 0x12, 0x45, 0x78, 0x65, 0x63, 0x75, 0x74, 0x65, 0x50, 0x6c, 0x61,
-	0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x20, 0x0a, 0x07, 0x67, 0x61, 0x6d, 0x65,
-	0x5f, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x42, 0x07, 0xba, 0x48, 0x04, 0x72, 0x02,
-	0x10, 0x01, 0x52, 0x06, 0x67, 0x61, 0x6d, 0x65, 0x49, 0x64, 0x12, 0x26, 0x0a, 0x09, 0x70, 0x69,
-	0x74, 0x5f, 0x69, 0x6e, 0x64, 0x65, 0x78, 0x18, 0x02, 0x20, 0x01, 0x28, 0x03, 0x42, 0x09, 0xba,
-	0x48, 0x06, 0x22, 0x04, 0x10, 0x06, 0x28, 0x00, 0x52, 0x08, 0x70, 0x69, 0x74, 0x49, 0x6e, 0x64,
-	0x65, 0x78, 0x22, 0x62, 0x0a, 0x13, 0x45, 0x78, 0x65, 0x63, 0x75, 0x74, 0x65, 0x50, 0x6c, 0x61,
-	0x79, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x28, 0x0a, 0x10, 0x70, 0x6c, 0x61,
-	0x79, 0x65, 0x64, 0x5f, 0x70, 0x69, 0x74, 0x5f, 0x69, 0x6e, 0x64, 0x65, 0x78, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x03, 0x52, 0x0e, 0x70, 0x6c, 0x61, 0x79, 0x65, 0x64, 0x50, 0x69, 0x74, 0x49, 0x6e,
-	0x64, 0x65, 0x78, 0x12, 0x21, 0x0a, 0x04, 0x67, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x0d, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x47, 0x61, 0x6d, 0x65,
-	0x52, 0x04, 0x67, 0x61, 0x6d, 0x65, 0x22, 0xcc, 0x01, 0x0a, 0x04, 0x47, 0x61, 0x6d, 0x65, 0x12,
-	0x0e, 0x0a, 0x02, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x02, 0x69, 0x64, 0x12,
-	0x21, 0x0a, 0x04, 0x74, 0x75, 0x72, 0x6e, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0e, 0x32, 0x0d, 0x2e,
-	0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x54, 0x75, 0x72, 0x6e, 0x52, 0x04, 0x74, 0x75,
-	0x72, 0x6e, 0x12, 0x27, 0x0a, 0x06, 0x72, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x18, 0x03, 0x20, 0x01,
-	0x28, 0x0e, 0x32, 0x0f, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x52, 0x65, 0x73,
-	0x75, 0x6c, 0x74, 0x52, 0x06, 0x72, 0x65, 0x73, 0x75, 0x6c, 0x74, 0x12, 0x33, 0x0a, 0x0b, 0x62,
-	0x6f, 0x61, 0x72, 0x64, 0x5f, 0x73, 0x69, 0x64, 0x65, 0x31, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x12, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x42, 0x6f, 0x61, 0x72, 0x64,
-	0x53, 0x69, 0x64, 0x65, 0x52, 0x0a, 0x62, 0x6f, 0x61, 0x72, 0x64, 0x53, 0x69, 0x64, 0x65, 0x31,
-	0x12, 0x33, 0x0a, 0x0b, 0x62, 0x6f, 0x61, 0x72, 0x64, 0x5f, 0x73, 0x69, 0x64, 0x65, 0x32, 0x18,
-	0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x12, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e,
-	0x42, 0x6f, 0x61, 0x72, 0x64, 0x53, 0x69, 0x64, 0x65, 0x52, 0x0a, 0x62, 0x6f, 0x61, 0x72, 0x64,
-	0x53, 0x69, 0x64, 0x65, 0x32, 0x22, 0x5e, 0x0a, 0x09, 0x42, 0x6f, 0x61, 0x72, 0x64, 0x53, 0x69,
-	0x64, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x70, 0x69, 0x74, 0x73, 0x18, 0x01, 0x20, 0x03, 0x28, 0x03,
-	0x52, 0x04, 0x70, 0x69, 0x74, 0x73, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x74, 0x6f, 0x72, 0x65, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x03, 0x52, 0x05, 0x73, 0x74, 0x6f, 0x72, 0x65, 0x12, 0x27, 0x0a, 0x06,
-	0x70, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x0f, 0x2e, 0x6d,
-	0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x50, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x52, 0x06, 0x70,
-	0x6c, 0x61, 0x79, 0x65, 0x72, 0x22, 0x32, 0x0a, 0x06, 0x50, 0x6c, 0x61, 0x79, 0x65, 0x72, 0x12,
-	0x12, 0x0a, 0x04, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x6e,
-	0x61, 0x6d, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x73, 0x63, 0x6f, 0x72, 0x65, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x03, 0x52, 0x05, 0x73, 0x63, 0x6f, 0x72, 0x65, 0x2a, 0x40, 0x0a, 0x04, 0x54, 0x75, 0x72,
-	0x6e, 0x12, 0x14, 0x0a, 0x10, 0x54, 0x55, 0x52, 0x4e, 0x5f, 0x55, 0x4e, 0x53, 0x50, 0x45, 0x43,
-	0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x00, 0x12, 0x10, 0x0a, 0x0c, 0x54, 0x55, 0x52, 0x4e, 0x5f,
-	0x50, 0x4c, 0x41, 0x59, 0x45, 0x52, 0x31, 0x10, 0x01, 0x12, 0x10, 0x0a, 0x0c, 0x54, 0x55, 0x52,
-	0x4e, 0x5f, 0x50, 0x4c, 0x41, 0x59, 0x45, 0x52, 0x32, 0x10, 0x02, 0x2a, 0x62, 0x0a, 0x06, 0x52,
-	0x65, 0x73, 0x75, 0x6c, 0x74, 0x12, 0x16, 0x0a, 0x12, 0x52, 0x45, 0x53, 0x55, 0x4c, 0x54, 0x5f,
-	0x55, 0x4e, 0x53, 0x50, 0x45, 0x43, 0x49, 0x46, 0x49, 0x45, 0x44, 0x10, 0x00, 0x12, 0x17, 0x0a,
-	0x13, 0x52, 0x45, 0x53, 0x55, 0x4c, 0x54, 0x5f, 0x50, 0x4c, 0x41, 0x59, 0x45, 0x52, 0x31, 0x5f,
-	0x57, 0x49, 0x4e, 0x53, 0x10, 0x01, 0x12, 0x17, 0x0a, 0x13, 0x52, 0x45, 0x53, 0x55, 0x4c, 0x54,
-	0x5f, 0x50, 0x4c, 0x41, 0x59, 0x45, 0x52, 0x32, 0x5f, 0x57, 0x49, 0x4e, 0x53, 0x10, 0x02, 0x12,
-	0x0e, 0x0a, 0x0a, 0x52, 0x45, 0x53, 0x55, 0x4c, 0x54, 0x5f, 0x54, 0x49, 0x45, 0x10, 0x03, 0x32,
-	0xdb, 0x01, 0x0a, 0x07, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12, 0x45, 0x0a, 0x0a, 0x43,
-	0x72, 0x65, 0x61, 0x74, 0x65, 0x47, 0x61, 0x6d, 0x65, 0x12, 0x1a, 0x2e, 0x6d, 0x61, 0x6e, 0x63,
-	0x61, 0x6c, 0x61, 0x2e, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x47, 0x61, 0x6d, 0x65, 0x52, 0x65,
-	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x1b, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e,
-	0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x47, 0x61, 0x6d, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e,
-	0x73, 0x65, 0x12, 0x3f, 0x0a, 0x08, 0x46, 0x69, 0x6e, 0x64, 0x47, 0x61, 0x6d, 0x65, 0x12, 0x18,
-	0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x46, 0x69, 0x6e, 0x64, 0x47, 0x61, 0x6d,
-	0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x19, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61,
-	0x6c, 0x61, 0x2e, 0x46, 0x69, 0x6e, 0x64, 0x47, 0x61, 0x6d, 0x65, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x12, 0x48, 0x0a, 0x0b, 0x45, 0x78, 0x65, 0x63, 0x75, 0x74, 0x65, 0x50, 0x6c,
-	0x61, 0x79, 0x12, 0x1b, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x45, 0x78, 0x65,
-	0x63, 0x75, 0x74, 0x65, 0x50, 0x6c, 0x61, 0x79, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x1a,
-	0x1c, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2e, 0x45, 0x78, 0x65, 0x63, 0x75, 0x74,
-	0x65, 0x50, 0x6c, 0x61, 0x79, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x77, 0x0a,
-	0x0b, 0x63, 0x6f, 0x6d, 0x2e, 0x6d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x42, 0x0c, 0x4d, 0x61,
-	0x6e, 0x63, 0x61, 0x6c, 0x61, 0x50, 0x72, 0x6f, 0x74, 0x6f, 0x50, 0x01, 0x5a, 0x1e, 0x67, 0x69,
-	0x74, 0x68, 0x75, 0x62, 0x2e, 0x63, 0x6f, 0x6d, 0x2f, 0x70, 0x63, 0x72, 0x69, 0x76, 0x2f, 0x6d,
-	0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0xa2, 0x02, 0x03, 0x4d,
-	0x58, 0x58, 0xaa, 0x02, 0x07, 0x4d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0xca, 0x02, 0x07, 0x4d,
-	0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0xe2, 0x02, 0x13, 0x4d, 0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61,
-	0x5c, 0x47, 0x50, 0x42, 0x4d, 0x65, 0x74, 0x61, 0x64, 0x61, 0x74, 0x61, 0xea, 0x02, 0x07, 0x4d,
-	0x61, 0x6e, 0x63, 0x61, 0x6c, 0x61, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_mancala_proto_rawDesc = "" +
+	"\n" +
+	"\rmancala.proto\x12\amancala\x1a\x1bbuf/validate/validate.proto\"Y\n" +
+	"\x11CreateGameRequest\x12!\n" +
+	"\aplayer1\x18\x01 \x01(\tB\a\xbaH\x04r\x02\x10\x01R\aplayer1\x12!\n" +
+	"\aplayer2\x18\x02 \x01(\tB\a\xbaH\x04r\x02\x10\x01R\aplayer2\"F\n" +
+	"\x12CreateGameResponse\x120\n" +
+	"\fcreated_game\x18\x01 \x01(\v2\r.mancala.GameR\vcreatedGame\"!\n" +
+	"\x0fFindGameRequest\x12\x0e\n" +
+	"\x02id\x18\x01 \x01(\tR\x02id\"5\n" +
+	"\x10FindGameResponse\x12!\n" +
+	"\x04game\x18\x01 \x01(\v2\r.mancala.GameR\x04game\"^\n" +
+	"\x12ExecutePlayRequest\x12 \n" +
+	"\agame_id\x18\x01 \x01(\tB\a\xbaH\x04r\x02\x10\x01R\x06gameId\x12&\n" +
+	"\tpit_index\x18\x02 \x01(\x03B\t\xbaH\x06\"\x04\x10\x06(\x00R\bpitIndex\"b\n" +
+	"\x13ExecutePlayResponse\x12(\n" +
+	"\x10played_pit_index\x18\x01 \x01(\x03R\x0eplayedPitIndex\x12!\n" +
+	"\x04game\x18\x02 \x01(\v2\r.mancala.GameR\x04game\"\xcc\x01\n" +
+	"\x04Game\x12\x0e\n" +
+	"\x02id\x18\x01 \x01(\tR\x02id\x12!\n" +
+	"\x04turn\x18\x02 \x01(\x0e2\r.mancala.TurnR\x04turn\x12'\n" +
+	"\x06result\x18\x03 \x01(\x0e2\x0f.mancala.ResultR\x06result\x123\n" +
+	"\vboard_side1\x18\x04 \x01(\v2\x12.mancala.BoardSideR\n" +
+	"boardSide1\x123\n" +
+	"\vboard_side2\x18\x05 \x01(\v2\x12.mancala.BoardSideR\n" +
+	"boardSide2\"^\n" +
+	"\tBoardSide\x12\x12\n" +
+	"\x04pits\x18\x01 \x03(\x03R\x04pits\x12\x14\n" +
+	"\x05store\x18\x02 \x01(\x03R\x05store\x12'\n" +
+	"\x06player\x18\x03 \x01(\v2\x0f.mancala.PlayerR\x06player\"2\n" +
+	"\x06Player\x12\x12\n" +
+	"\x04name\x18\x01 \x01(\tR\x04name\x12\x14\n" +
+	"\x05score\x18\x02 \x01(\x03R\x05score*@\n" +
+	"\x04Turn\x12\x14\n" +
+	"\x10TURN_UNSPECIFIED\x10\x00\x12\x10\n" +
+	"\fTURN_PLAYER1\x10\x01\x12\x10\n" +
+	"\fTURN_PLAYER2\x10\x02*b\n" +
+	"\x06Result\x12\x16\n" +
+	"\x12RESULT_UNSPECIFIED\x10\x00\x12\x17\n" +
+	"\x13RESULT_PLAYER1_WINS\x10\x01\x12\x17\n" +
+	"\x13RESULT_PLAYER2_WINS\x10\x02\x12\x0e\n" +
+	"\n" +
+	"RESULT_TIE\x10\x032\xdb\x01\n" +
+	"\aService\x12E\n" +
+	"\n" +
+	"CreateGame\x12\x1a.mancala.CreateGameRequest\x1a\x1b.mancala.CreateGameResponse\x12?\n" +
+	"\bFindGame\x12\x18.mancala.FindGameRequest\x1a\x19.mancala.FindGameResponse\x12H\n" +
+	"\vExecutePlay\x12\x1b.mancala.ExecutePlayRequest\x1a\x1c.mancala.ExecutePlayResponseBw\n" +
+	"\vcom.mancalaB\fMancalaProtoP\x01Z\x1egithub.com/pcriv/mancala/proto\xa2\x02\x03MXX\xaa\x02\aMancala\xca\x02\aMancala\xe2\x02\x13Mancala\\GPBMetadata\xea\x02\aMancalab\x06proto3"
 
 var (
 	file_mancala_proto_rawDescOnce sync.Once
-	file_mancala_proto_rawDescData = file_mancala_proto_rawDesc
+	file_mancala_proto_rawDescData []byte
 )
 
 func file_mancala_proto_rawDescGZIP() []byte {
 	file_mancala_proto_rawDescOnce.Do(func() {
-		file_mancala_proto_rawDescData = protoimpl.X.CompressGZIP(file_mancala_proto_rawDescData)
+		file_mancala_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_mancala_proto_rawDesc), len(file_mancala_proto_rawDesc)))
 	})
 	return file_mancala_proto_rawDescData
 }
@@ -750,7 +712,7 @@ func file_mancala_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_mancala_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_mancala_proto_rawDesc), len(file_mancala_proto_rawDesc)),
 			NumEnums:      2,
 			NumMessages:   9,
 			NumExtensions: 0,
@@ -762,7 +724,6 @@ func file_mancala_proto_init() {
 		MessageInfos:      file_mancala_proto_msgTypes,
 	}.Build()
 	File_mancala_proto = out.File
-	file_mancala_proto_rawDesc = nil
 	file_mancala_proto_goTypes = nil
 	file_mancala_proto_depIdxs = nil
 }
