@@ -1,6 +1,6 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.5.1
+// - protoc-gen-go-grpc v1.6.1
 // - protoc             (unknown)
 // source: mancala.proto
 
@@ -91,13 +91,13 @@ type ServiceServer interface {
 type UnimplementedServiceServer struct{}
 
 func (UnimplementedServiceServer) CreateGame(context.Context, *CreateGameRequest) (*CreateGameResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method CreateGame not implemented")
+	return nil, status.Error(codes.Unimplemented, "method CreateGame not implemented")
 }
 func (UnimplementedServiceServer) FindGame(context.Context, *FindGameRequest) (*FindGameResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method FindGame not implemented")
+	return nil, status.Error(codes.Unimplemented, "method FindGame not implemented")
 }
 func (UnimplementedServiceServer) ExecutePlay(context.Context, *ExecutePlayRequest) (*ExecutePlayResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method ExecutePlay not implemented")
+	return nil, status.Error(codes.Unimplemented, "method ExecutePlay not implemented")
 }
 func (UnimplementedServiceServer) testEmbeddedByValue() {}
 
@@ -109,7 +109,7 @@ type UnsafeServiceServer interface {
 }
 
 func RegisterServiceServer(s grpc.ServiceRegistrar, srv ServiceServer) {
-	// If the following call pancis, it indicates UnimplementedServiceServer was
+	// If the following call panics, it indicates UnimplementedServiceServer was
 	// embedded by pointer and is nil.  This will cause panics if an
 	// unimplemented method is ever invoked, so we test this at initialization
 	// time to prevent it from happening at runtime later due to I/O.
